Add: set icon for tool button with QIcon::fromtheme Add: small ui tweaks for GroupWidget

+++ widgets/ProjectMode/GroupWidget.cpp
@@ -4,6 +4,7 @@
 #include "../../core/SqlErrorException.h"
 #include "NavigationBarWidget.h"
 #include <QDebug>
+#include <QAction>
 #include <QMessageBox>
 
 GroupWidget::GroupWidget(QWidget *parent) :
@@ -12,38 +13,10 @@ GroupWidget::GroupWidget(QWidget *parent) :
 {
     ui->setupUi(this);
 
-    this->newItemInputDialog = new QInputDialog(this);
-
-    this->groupDetailWidget = new GroupDetailWidget(this);
-    ui->stackedWidget->addWidget(this->groupDetailWidget);
-    this->itemDetailWidget = new ItemDetailWidget(this);
-    ui->stackedWidget->addWidget(this->itemDetailWidget);
-    ui->stackedWidget->setCurrentWidget(this->groupDetailWidget);
-
-    this->listWidget = new ItemListWidget(ui->leftWidget);
-    ui->leftVLayout->addWidget(this->listWidget);
-
-    this->mainSplitter = new QSplitter(this);
-    this->mainSplitter->addWidget(ui->leftWidget);
-    this->mainSplitter->addWidget(ui->stackedWidget);
-    this->mainSplitter->setStretchFactor(0, 1);
-    this->mainSplitter->setStretchFactor(1, 3);
-    ui->horizontalLayout->addWidget(this->mainSplitter);
-
-    connect(this->listWidget, &ItemListWidget::selectedItemChanged, this, &GroupWidget::selected_item_changed);
-    connect(this->groupDetailWidget, &GroupDetailWidget::itemModified, this, &GroupWidget::current_item_modified);
-    connect(this->listWidget, &ItemListWidget::doubleClicked, this, &GroupWidget::item_double_clicked);
-
-    connect(ui->addItemGroupToolButton, &QToolButton::clicked, this, &GroupWidget::new_group_button_clicked);
-    connect(ui->addItemDetailToolButton, &QToolButton::clicked, this, &GroupWidget::new_detail_button_clicked);
-    connect(this->groupDetailWidget, &GroupDetailWidget::deleteButtonClicked, this, &GroupWidget::delete_group_button_clicked);
-
-    connect(&this->dataCenter, &todo::DataCenter::databaseModified, this, &GroupWidget::database_modified);
-    connect(this->groupDetailWidget, &GroupDetailWidget::saveActionTriggered, this, &GroupWidget::save_action_triggered);
-    connect(this->itemDetailWidget, &ItemDetailWidget::saveButtonClicked, this, &GroupWidget::save_action_triggered);
-
+    this->initUI();
     this->loadItems(this->getRootGroups());
     this->currPathList.append(NavigationBarWidget::ROOT);
+    this->addNewPopupMenu->setEnabled(false);
 }
 
 GroupWidget::~GroupWidget()
@@ -52,6 +25,13 @@ GroupWidget::~GroupWidget()
 }
 
 void GroupWidget::selected_item_changed(const QString &itemID) {
+    if (this->groupDetailWidget->isEditing()) {
+        if (QMessageBox::warning(this, tr("Attention !!"), tr("Your work not saved !. Save it ?"),
+                             QMessageBox::Yes | QMessageBox::No) == QMessageBox::Yes) {
+            this->save_action_triggered(this->groupDetailWidget->collectData());
+        }
+    }
+
     if (this->itemMap[itemID].isGroup()) {
         ui->stackedWidget->setCurrentWidget(this->groupDetailWidget);
         this->groupDetailWidget->loadItemGroup(this->itemMap[itemID].getItemGroup());
@@ -81,6 +61,7 @@ void GroupWidget::item_double_clicked(const QString &itemID) {
         emit enterItem(itemID, this->itemMap[itemID].getItemGroup().getTitle());
         this->loadItems(subItemList);
         this->currPathList.append(itemID);
+        this->addNewPopupMenu->setEnabled(true);
     }
 }
 
@@ -101,9 +82,10 @@ void GroupWidget::jump_to(const QList<QString> &pathList) {
     QString clickedID = pathList.last();
     if (clickedID != NavigationBarWidget::ROOT) {
         this->loadItems(this->getSubItemsForGroup(clickedID));
+        this->addNewPopupMenu->setEnabled(true);
     } else {
         this->loadItems(this->getRootGroups());
-
+        this->addNewPopupMenu->setEnabled(false);
     }
     this->currPathList = pathList;  // restore the full path
 }
@@ -138,6 +120,7 @@ void GroupWidget::dealWithNewItem(const todo::ItemAndGroupWrapper &newWrapper) {
 }
 
 void GroupWidget::new_group_button_clicked() {
+    this->newItemInputDialog->setLabelText("Please input the new project's title: ");
     this->newItemInputDialog->setTextValue(tr(""));
     int ok = this->newItemInputDialog->exec();
     if (ok) {
@@ -159,6 +142,7 @@ void GroupWidget::new_group_button_clicked() {
 }
 
 void GroupWidget::new_detail_button_clicked() {
+    this->newItemInputDialog->setLabelText("Please input the new task's title: ");
     this->newItemInputDialog->setTextValue(tr(""));
     int ok = this->newItemInputDialog->exec();
     if (ok) {
@@ -174,10 +158,13 @@ void GroupWidget::new_detail_button_clicked() {
 }
 
 void GroupWidget::delete_group_button_clicked() {
-    QString itemID = this->listWidget->currentItemID();
-    this->dataCenter.deleteGroupCompletely(itemID);
-    this->listWidget->removeItemWrapperByID(itemID);
-    this->itemMap.remove(itemID);
+    if (QMessageBox::warning(this, tr("Attention !!"), tr("You are trying to delete it. Sure to delete it ?"),
+                             QMessageBox::Yes | QMessageBox::No) == QMessageBox::Yes) {
+        QString itemID = this->listWidget->currentItemID();
+        this->dataCenter.deleteGroupCompletely(itemID);
+        this->listWidget->removeItemWrapperByID(itemID);
+        this->itemMap.remove(itemID);
+    }
 }
 
 void GroupWidget::database_modified() {
@@ -198,4 +185,55 @@ void GroupWidget::save_action_triggered(const todo::ItemAndGroupWrapper &wrapper
         this->itemMap[detail.getId()] = detail;
         this->listWidget->refresh_item_info(detail);
     }
+}
+
+void GroupWidget::refresh_current_items() {
+    QString currItemID = this->currPathList.last();
+    if (currItemID == NavigationBarWidget::ROOT) {
+        this->loadItems(this->getRootGroups());
+    } else {
+        this->loadItems(this->getSubItemsForGroup(currItemID));
+    }
+}
+
+void GroupWidget::initUI() {
+    this->newItemInputDialog = new QInputDialog(this);
+
+    this->groupDetailWidget = new GroupDetailWidget(this);
+    ui->stackedWidget->addWidget(this->groupDetailWidget);
+    this->itemDetailWidget = new ItemDetailWidget(this);
+    ui->stackedWidget->addWidget(this->itemDetailWidget);
+    ui->stackedWidget->setCurrentWidget(this->groupDetailWidget);
+
+    this->listWidget = new ItemListWidget(ui->leftWidget);
+    ui->leftVLayout->addWidget(this->listWidget);
+
+    this->mainSplitter = new QSplitter(this);
+    this->mainSplitter->addWidget(ui->leftWidget);
+    this->mainSplitter->addWidget(ui->stackedWidget);
+    this->mainSplitter->setStretchFactor(0, 1);
+    this->mainSplitter->setStretchFactor(1, 3);
+    ui->horizontalLayout->addWidget(this->mainSplitter);
+
+    this->addNewPopupMenu = new QMenu(ui->addToolButton);
+    ui->addToolButton->setMenu(this->addNewPopupMenu);
+    QAction *addNewItemAction = new QAction("Add new task", this->addNewPopupMenu);
+    connect(addNewItemAction, &QAction::triggered, this, &GroupWidget::new_detail_button_clicked);
+    this->addNewPopupMenu->addAction(addNewItemAction);
+
+    connect(this->listWidget, &ItemListWidget::selectedItemChanged, this, &GroupWidget::selected_item_changed);
+    connect(this->groupDetailWidget, &GroupDetailWidget::itemModified, this, &GroupWidget::current_item_modified);
+    connect(this->listWidget, &ItemListWidget::doubleClicked, this, &GroupWidget::item_double_clicked);
+
+    connect(ui->addToolButton, &QToolButton::clicked, this, &GroupWidget::new_group_button_clicked);
+    connect(this->groupDetailWidget, &GroupDetailWidget::deleteButtonClicked, this, &GroupWidget::delete_group_button_clicked);
+
+    connect(&this->dataCenter, &todo::DataCenter::databaseModified, this, &GroupWidget::database_modified);
+    connect(this->groupDetailWidget, &GroupDetailWidget::saveActionTriggered, this, &GroupWidget::save_action_triggered);
+    connect(this->itemDetailWidget, &ItemDetailWidget::saveButtonClicked, this, &GroupWidget::save_action_triggered);
+
+    // set icons
+    ui->refreshToolButton->setIcon(QIcon::fromTheme("view-refresh"));
+    ui->sortToolButton->setIcon(QIcon::fromTheme("sort-presence"));
+    ui->addToolButton->setIcon(QIcon::fromTheme("list-add"));
 }
